requirements.txt
astroid==2.0.4
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
defusedxml==0.5.0
Django==2.0
django-allauth==0.37.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.9
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
Pillow==5.2.0
psycopg2-binary==2.7.5
pylint==2.1.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
whitenoise==4.0
wrapt==1.10.11
